+++ packages/components/src/Card/CardClickable.tsx
@@ -5,6 +5,8 @@ interface ClickableCardProps {
   onClick(event: React.MouseEvent<HTMLElement>): void;
   readonly className: string;
   readonly children: ReactNode | ReactNode[];
+  readonly UNSAFE_className?: string;
+  readonly UNSAFE_style?: React.CSSProperties;
 }
 
 const ENTER_KEY = "Enter";
@@ -18,6 +20,7 @@ export function CardClickable({
   className,
   onClick,
   children,
+  UNSAFE_style = {},
 }: ClickableCardProps) {
   const cardRef = useRef() as React.MutableRefObject<HTMLDivElement>;
 
@@ -26,6 +29,7 @@ export function CardClickable({
       ref={cardRef}
       data-testid="clickable-card"
       className={className}
+      style={UNSAFE_style}
       onClick={onClick}
       onKeyUp={handleKeyup}
       onKeyDown={handleKeyDown}